[refactor]: useModal 컴포넌트 리턴 제거(#126)
