+++ src/hooks/useModal.ts
@@ -1,0 +1,18 @@
+import { useState } from 'react';
+
+export const useModal = () => {
+  const [isOpen, setIsOpen] = useState(false);
+
+  // 배경 스크롤링 방지
+  // 필요시 훅으로 분리
+  if (isOpen) {
+    document.body.style.overflow = 'hidden';
+  } else {
+    document.body.style.overflow = 'auto';
+  }
+
+  const open = () => setIsOpen(true);
+  const close = () => setIsOpen(false);
+
+  return { isOpen, open, close };
+};
